backend/requirements.txt
alembic==1.7.7
APScheduler==3.9.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
blinker==1.4
click==8.1.3
Deprecated==1.2.13
Flask==2.1.2
Flask-APScheduler==1.12.3
Flask-Cors==3.0.10
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-redis==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
ImageHash==4.2.1
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.22.4
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
PyJWT==2.4.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
redis==4.3.1
scipy==1.8.1
six==1.16.0
SQLAlchemy==1.4.36
tzdata==2022.1
tzlocal==4.2
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
